refactor to plugin architecture
