+++ ferry/lib/cache.dart
@@ -1,8 +1,6 @@
 export 'package:ferry_store/ferry_store.dart';
 
 export './src/cache.dart';
-export './src/cache_proxy.dart';
-export './src/client_options.dart';
 export './src/fetch_policy.dart';
 export './src/operation_request.dart';
 export './src/fragment_request.dart';
