+++ ferry/lib/ferry.dart
@@ -2,8 +2,6 @@ export 'package:gql_link/gql_link.dart';
 export 'package:ferry_store/ferry_store.dart';
 
 export './src/cache.dart';
-export './src/cache_proxy.dart';
-export './src/client_options.dart';
 export './src/client.dart';
 export './src/fetch_policy.dart';
 export './src/operation_request.dart';
