+++ ferry/lib/plugins.dart
@@ -1,0 +1,4 @@
+export './src/plugins/plugin.dart';
+export './src/plugins/add_typename_plugin/add_typename_plugin.dart';
+export './src/plugins/update_cache_plugin/update_cache_plugin.dart';
+export './src/plugins/update_result_plugin/update_result_plugin.dart';
