+++ ferry/test/multiple_requests_test.dart
@@ -3,6 +3,7 @@ import 'package:mockito/mockito.dart';
 import 'package:gql_link/gql_link.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:ferry/ferry.dart';
+import 'package:ferry/plugins.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:test/test.dart';
 
@@ -22,8 +23,7 @@ void main() {
 
     final client = Client(
       link: mockLink,
-      options: ClientOptions(addTypename: false),
-    );
+    )..plugins.removeWhere((plugin) => plugin is AddTypenamePlugin);
     test(
         "Requests with different args aren't returned in the same response stream",
         () async {
