+++ ferry/test/plugins/update_cache_plugin_test.dart
@@ -4,6 +4,7 @@ import 'package:mockito/mockito.dart';
 import 'package:gql_link/gql_link.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:ferry/ferry.dart';
+import 'package:ferry/plugins.dart';
 import 'package:test/test.dart';
 
 import 'package:ferry_test_graphql/queries/variables/reviews.req.gql.dart';
@@ -12,6 +13,9 @@ import 'package:ferry_test_graphql/mutations/variables/create_review.req.gql.dar
 import 'package:ferry_test_graphql/mutations/variables/create_review.var.gql.dart';
 import 'package:ferry_test_graphql/mutations/variables/create_review.data.gql.dart';
 import 'package:ferry_test_graphql/schema/schema.schema.gql.dart';
+
+import 'package:ferry_test_graphql/queries/variables/human_with_args.req.gql.dart';
+import 'package:ferry_test_graphql/queries/variables/human_with_args.data.gql.dart';
 
 class MockLink extends Mock implements Link {}
 
@@ -39,11 +43,27 @@ UpdateCacheHandler<GCreateReviewData, GCreateReviewVars> createReviewHandler = (
 };
 
 void main() {
-  group('UpdateCacheHandler', () {
-    final clientOptions = ClientOptions(updateCacheHandlers: {
-      'createReviewHandler': createReviewHandler,
-    });
+  group('CacheProxy', () {
+    group('writing queries', () {
+      final cache = Cache();
 
+      final req = GHumanWithArgsReq((b) => b..vars.id = '123');
+      final data = GHumanWithArgsData(
+        (b) => b
+          ..human.name = 'Han Solo'
+          ..human.height = 1.85,
+      );
+
+      final cacheProxy = CacheProxy(cache, true, req.requestId);
+      test('can write queries', () {
+        expect(cacheProxy.readQuery(req), equals(null));
+        cacheProxy.writeQuery(req, data);
+        expect(cacheProxy.readQuery(req), equals(data));
+      });
+    });
+  });
+
+  group('UpdateCacheHandler', () {
     group('without optimistic response', () {
       final mockLink = MockLink();
       final req = GCreateReviewReq(
@@ -54,7 +74,12 @@ void main() {
           ..vars.review.commentary = 'Amazing!!!',
       );
       final linkController = StreamController<Response>();
-      final client = Client(link: mockLink, options: clientOptions);
+      final client = Client(link: mockLink);
+      final updateCachePlugin =
+          UpdateCachePlugin(cache: client.cache, updateCacheHandlers: {
+        'createReviewHandler': createReviewHandler,
+      });
+      client.plugins.add(updateCachePlugin);
       final queue = StreamQueue(client.responseStream(
         req,
         executeOnListen: false,
@@ -92,7 +117,12 @@ void main() {
           ..optimisticResponse.createReview.commentary = 'hi',
       );
       final linkController = StreamController<Response>();
-      final client = Client(link: mockLink, options: clientOptions);
+      final client = Client(link: mockLink);
+      final updateCachePlugin =
+          UpdateCachePlugin(cache: client.cache, updateCacheHandlers: {
+        'createReviewHandler': createReviewHandler,
+      });
+      client.plugins.add(updateCachePlugin);
       final queue = StreamQueue(client.responseStream(
         req,
         executeOnListen: false,
